requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.3
django-datetime-widget==0.9.3
django-filter==1.1.0
django-formtools==2.1
django-rest-framework==0.1.0
djangorestframework==3.7.7
drfapikey==0.0.3
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
